pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
coloraide = "^3.3.1"
numpy = "^1.26.4"
pillow = "^10.3.0"
pyfastnoisesimd = "^0.4.2"
scikit-image = "^0.23.2"
gdal = "^3.8.5"
scipy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
